try workaround bazel issue
